third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/gelu.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] gelu float16 0D scalar
  assert_less_than_equal: assert_array_approx_equals_ulp: test gelu float16 actual 0 should be close enough to expected -0.021636962890625 by ULP distance: expected a number less than or equal to 18 but got 42378
[FAIL] gelu float16 1D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test gelu float16 actual 0 should be close enough to expected 0.71142578125 by ULP distance: expected a number less than or equal to 18 but got 14769
[FAIL] gelu float16 2D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test gelu float16 actual 0 should be close enough to expected 0.71142578125 by ULP distance: expected a number less than or equal to 18 but got 14769
[FAIL] gelu float16 3D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test gelu float16 actual 0 should be close enough to expected 0.71142578125 by ULP distance: expected a number less than or equal to 18 but got 14769
[FAIL] gelu float16 4D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test gelu float16 actual 0 should be close enough to expected 0.71142578125 by ULP distance: expected a number less than or equal to 18 but got 14769
Harness: the test ran to completion.
